Add advanced_vpn_settings variable attribute
Following this page led to a terraform error for a missing "advanced_vpn_settings" attribute, which is a recently added attribute to the variables.tf of the Azure Landing Zone module. 

Updated settings.connectivity.tf example in this documentation page to include advanced_vpn_settings, as per the default values specified in variables.tf of the module. This resolves the TF error.